backend/requirements/requirements-base.txt
Django>=3.2.13,<4
django-filter==21.1
djangorestframework==3.13.1
django-cors-headers==3.12.0
drf-spectacular==0.22.1
Markdown==3.3.7
python-dotenv==0.20.0
drf-nested-routers==0.93.4
django-model-utils==4.2.0
djangorestframework-simplejwt==4.8.0
dj-rest-auth==2.2.5
Pillow==9.2.0
python-decouple==3.6
drf-extensions==0.7.1
djangorestframework-camel-case==1.3.0

# DI
# =====================
dependency-injector==4.39.1
